+++ Tests/ValidationRules.Replication.Tests/ValidationRules.Replication.Tests.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -190,6 +190,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BindingObjectSpecificationTests.cs" />
+    <Compile Include="CouponBeginEndMonthTests.cs" />
     <Compile Include="DataFixtureBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
